pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coding-ai-agent"
version = "0.1.0"
description = "Coding AI Agent built on DeepAgents with long-term memory, dynamic sub-agents, and model fallback"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-ollama>=0.3.0",
    # Memory
    "chromadb>=1.0.0",
    # WebUI
    "streamlit>=1.40.0",
    # Utils
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
deepagents = [
    "deepagents @ git+https://github.com/langchain-ai/deepagents.git#subdirectory=libs/deepagents",
]

[project.scripts]
coding-agent = "coding_agent.__main__:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/coding_agent"]
